Fixed Sending bug on invite modal
